+++ app/api/api/views.py
@@ -76,9 +76,9 @@ from shared.services.rules_export import (
     get_mapping_rules_list,
     make_dag,
 )
-from shared_code import storage_router
+from shared.services.storage_router import StorageService
 
-storage_parser = storage_router.StorageService()
+storage_parser = StorageService()
 
 
 class DataPartnerViewSet(GenericAPIView, ListModelMixin):
