+++ app/shared/shared/mapping/views.py
@@ -39,12 +39,12 @@ from shared.services.rules_export import (
     get_mapping_rules_json,
     make_dag,
 )
-from shared_code import storage_router
+from shared.services.storage_router import StorageService
 
 from .forms import ScanReportAssertionForm, ScanReportForm
 from .permissions import has_editorship, has_viewership, is_admin
 
-storage_parser = storage_router.StorageService()
+storage_parser = StorageService()
 
 
 @login_required
