+++ app/shared/shared/services/utils.py
