+++ app/workers/test/shared_code/test_blob_parser.py
@@ -1,5 +1,5 @@
 import pytest
-from shared_code.utils import (
+from shared.services.utils import (
     process_four_item_dict,
     process_three_item_dict,
     remove_BOM,
